+++ app/src/main/java/com/teampatch/harmony/MainNavHost.kt
@@ -16,6 +16,8 @@ import com.teampatch.feature.home.addHomeScreen
 import com.teampatch.feature.home.navigateToHomeScreen
 import com.teampatch.feature.memorycard.registration.addMemoryCardRegistrationScreen
 import com.teampatch.feature.memorycard.registration.navigateToMemoryCardRegistrationScreen
+import com.teampatch.feature.onboarding.enter.addOnboardingEnterInvitationCodeScreen
+import com.teampatch.feature.onboarding.enter.navigateToEnterInvitationCodeScreen
 import com.teampatch.feature.onboarding.login.ui.OnboardingRoute
 import com.teampatch.feature.onboarding.login.ui.addOnboardingEnterInvitationCodeScreen
 import com.teampatch.feature.onboarding.login.ui.addOnboardingEnterSpaceScreen
@@ -92,7 +94,8 @@ fun MainNavHost(
         )
 
         addOnboardingEnterInvitationCodeScreen(
-            onNextClick = { navController.navigateToEnterSpaceScreen() }
+            onBackRequest = navController::popBackStack,
+            onEnterSpaceScreenRequest = { navController.navigateToEnterSpaceScreen() }
         )
 
         addOnboardingEnterSpaceScreen(onNextClick = { navController.navigateToHomeScreen() })
