--- feature/onboarding/src/main/java/com/teampatch/feature/onboarding/login/ui/OnboardingMakeGroupScreen.java
@@ -1,5 +1,0 @@
-package com.teampatch.feature.onboarding.login.ui;
-
-public class OnboardingMakeGroupScreen {
-    // Figma 완성이 안되어서 비워놓음
-}
