+++ feature/onboarding/src/main/java/com/teampatch/feature/onboarding/login/ui/OnboardingNavigation.kt
@@ -75,43 +75,3 @@ fun NavGraphBuilder.addOnboardingStartScreen(
         )
     }
 }
-
-@Serializable
-data object OnboardingEnterInvitationCodeRoute
-
-fun NavController.navigateToEnterInvitationCodeScreen(
-    navOptions: NavOptions? = null,
-    navigatorExtras: Navigator.Extras? = null,
-) {
-    navigate(OnboardingEnterInvitationCodeRoute, navOptions, navigatorExtras)
-}
-
-fun NavGraphBuilder.addOnboardingEnterInvitationCodeScreen(
-    onNextClick: () -> Unit,
-) {
-    composable<OnboardingEnterInvitationCodeRoute> {
-        OnboardingEnterInvitationCodeScreen(
-            onNextClick = onNextClick
-        )
-    }
-}
-
-@Serializable
-data object OnboardingEnterSpaceRoute
-
-fun NavController.navigateToEnterSpaceScreen(
-    navOptions: NavOptions? = null,
-    navigatorExtras: Navigator.Extras? = null,
-) {
-    navigate(OnboardingEnterSpaceRoute, navOptions, navigatorExtras)
-}
-
-fun NavGraphBuilder.addOnboardingEnterSpaceScreen(
-    onNextClick: () -> Unit,
-) {
-    composable<OnboardingEnterSpaceRoute> {
-        OnboardingEnterSpaceScreen(
-            onNextClick = onNextClick
-        )
-    }
-}
